Make the app default to its own settings
This sets the prefix correctly and auto-generates a game ID. What remains
is to enter the opponents' prefix an their game ID, and then the game
can set itself up correctly.

Pressing the button works, but it's all done on the main thread, which is why it
is so slow.