Implement the grid to work with the pings
This actually colors the grid in the way the remote is. Now we need to add a local / remote dual grid thing and after that, indicate which positions the other player chose. Let's hope we can actually use ip6fw for that, otherwise we have to do another nice tcpdump hack.
 